Generate metadata in one place
